(PC-26282)[PRO] refactor: move select handler in its component
